implementation-2/requirements.txt
streamlit 
pillow
numpy
tensorflow
tensorflow-hub 
matplotlib
torch
torchvision
Pillow